refactor(dma): refactor DMA and SDIO implementation with ownership and API improvements
