+++ bouffalo-hal/src/dma/mod.rs
@@ -43,8 +43,32 @@ pub enum DmaAddr {
 
 /// Extend constructor to DMA ownership structures.
 pub trait DmaExt {
-    type Group<'a>
-    where
-        Self: 'a;
-    fn split(&self, glb: &glb::v2::RegisterBlock) -> Self::Group<'_>;
+    type Group;
+    fn split(self, glb: &glb::v2::RegisterBlock) -> Self::Group;
+}
+
+/// Linked list item pool descriptor.
+#[derive(Clone, Copy, Debug, PartialEq, Eq, Hash)]
+#[repr(C)]
+pub struct LliPool {
+    /// Source address.
+    pub src_addr: u32,
+    /// Destination address.
+    pub dst_addr: u32,
+    /// Physical address to next linked list item.
+    pub next_lli: u32,
+    /// Linked list item control register.
+    pub control: LliControl,
+}
+
+impl LliPool {
+    #[inline]
+    pub fn new() -> Self {
+        Self {
+            src_addr: 0,
+            dst_addr: 0,
+            next_lli: 0,
+            control: LliControl::default(),
+        }
+    }
 }
