+++ bouffalo-rt/src/macros.rs
@@ -2,8 +2,7 @@ macro_rules! soc {
     (
         $(
             $(#[$doc:meta])*
-            pub struct $Ty:ident => $paddr:expr_2021
-            $(, $DerefTy:ty)+ ;
+            pub struct $Ty:ident => $paddr:expr_2021, $DerefTy:ty;
         )+
     ) => {
         $(
@@ -13,23 +12,26 @@ macro_rules! soc {
                 _private: (),
             }
 
-            $(
-                impl core::ops::Deref for $Ty {
-                    type Target = $DerefTy;
-
-                    #[inline(always)]
-                    fn deref(&self) -> &Self::Target {
-                        unsafe { &*($paddr as *const _) }
-                    }
+            impl $Ty {
+                #[inline]
+                pub const fn as_ptr(self) -> *const $DerefTy {
+                    $paddr as *const _
                 }
+            }
 
-                impl core::convert::AsRef<$DerefTy> for $Ty {
-                    #[inline(always)]
-                    fn as_ref(&self) -> &$DerefTy {
-                        unsafe { &*($paddr as *const _) }
-                    }
+            impl core::ops::Deref for $Ty {
+                type Target = $DerefTy;
+                #[inline(always)]
+                fn deref(&self) -> &Self::Target {
+                    unsafe { &*($paddr as *const _) }
                 }
-            )+
+            }
+            impl core::convert::AsRef<$DerefTy> for $Ty {
+                #[inline(always)]
+                fn as_ref(&self) -> &$DerefTy {
+                    unsafe { &*($paddr as *const _) }
+                }
+            }
         )+
     };
 }
